cli for base58 peer-id to hex complete. It produces the bytes(Vec<u8>) and the hex when provided peerid
