+++ Cargo.toml
@@ -5,7 +5,6 @@ members = [
     'node',
     'pallets/*',
     'runtime',
-    'supra-cli',
 ]
 [profile.release]
 panic = 'unwind'
