+++ node/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod chain_spec;
 pub mod rpc;
 pub mod service;
+pub mod decode_peer_id;
