--- supra-cli/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "supra-cli"
-version = "0.1.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-bs58 = "0.4.0"
-hex = "0.4.3"
-sc-cli = "0.9.0"
-structopt = "0.3.23"
-libp2p = "0.34.0"
-
